requirements.txt
streamlit
python-dotenv
PyPDF2
pinecone
google-genai
